+++ lib/src/menu/widgets/button.dart
@@ -1,5 +1,7 @@
 /// Provides the [Button] class.
 import '../../json/asset_reference.dart';
+import '../../json/message.dart';
+import '../menu_base.dart';
 import 'widgets_base.dart';
 
 /// A button that can be activated.
@@ -43,6 +45,17 @@ class ListButton<T> extends Widget {
 
   /// Get the currently-focused item.
   T get value => items[index];
+
+  /// Get a label that shows the current state of this button.
+  @override
+  Message getLabel(MenuItem menuItem) {
+    final label = menuItem.label;
+    return Message(
+        text: '${label.text} ($value)',
+        gain: label.gain,
+        keepAlive: label.keepAlive,
+        sound: label.sound);
+  }
 }
 
 /// A checkbox.
@@ -50,6 +63,40 @@ class ListButton<T> extends Widget {
 /// This control can toggle between `true` and `false`.
 class Checkbox extends ListButton<bool> {
   /// Create an instance.
-  Checkbox(void Function(bool) onChange, {bool initialValue = true})
+  Checkbox(void Function(bool) onChange,
+      {bool initialValue = true, this.checkedSound, this.uncheckedSound})
       : super([initialValue, !initialValue], onChange);
+
+  /// The sound that will be played when this checkbox is selected if [value] is
+  /// `true`.
+  ///
+  /// If this value is `null`, the sound will default to the sound of the
+  /// [MenuItem] label for this checkbox. If no sound is desired, ensure the
+  /// menu item sound is also `null`.
+  final AssetReference? checkedSound;
+
+  /// The sound that will be played when this checkbox is selected if [value] is
+  /// `false`.
+  ///
+  /// If this value is `null`, the sound will default to the sound of the
+  /// [MenuItem] label for this checkbox. If no sound is desired, ensure the
+  /// menu item sound is also `null`.
+  final AssetReference? uncheckedSound;
+
+  @override
+  Message getLabel(MenuItem menuItem) {
+    final label = menuItem.label;
+    AssetReference? sound;
+    if (value) {
+      sound = checkedSound;
+    } else {
+      sound = uncheckedSound;
+    }
+    sound ??= label.sound;
+    return Message(
+        text: '${label.text} (${value == true ? "checked" : "unchecked"})',
+        gain: label.gain,
+        keepAlive: label.keepAlive,
+        sound: sound);
+  }
 }