+++ lib/src/menu/widgets/widgets_base.dart
@@ -1,7 +1,9 @@
 /// Widgets for use with [Menu] instances.
 library widgets;
 
+import '../../json/message.dart';
 import '../menu.dart';
+import '../menu_item.dart';
 
 export 'button.dart';
 export 'label.dart';
@@ -10,4 +12,8 @@ export 'label.dart';
 class Widget {
   /// Allow subclasses to be constant.
   const Widget();
+
+  /// Override this method to change the label that is shown when this widget's
+  /// parent [MenuItem] is selected.
+  Message? getLabel(MenuItem menuItem) {}
 }